% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auc_toy_vector}
\alias{auc_toy_vector}
\title{AUC toy example}
\format{
An object of class \code{numeric} of length 81068.
}
\usage{
data(auc_toy_vector)
}
\description{
A toy example of AUC vector produced by the compute_auc() function.
}
\keyword{datasets}
