% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AUC.R
\name{compute_AUC}
\alias{compute_AUC}
\title{Compute Area Under Curve for a matrix of samples}
\usage{
compute_AUC(tumor_table, control_table, ncores = 1, na_threshold = 0,
  simplify = TRUE)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (percentage) from tumor samples.}

\item{control_table}{A matrix of beta-values (percentage) from normal/control samples.}

\item{ncores}{Number of parallel processes to use for parallel computing.}

\item{na_threshold}{Fraction of NAs (considered independently in tumor and
control samples) above which a site will not be selected (default=0).}

\item{simplify}{If TRUE (default) return a vector else compute all AUC and return a data.frame
reporting fraction of NAs in tumor and control tables.}
}
\value{
A vector of AUC scores.
}
\description{
This function computes the Area Under Curve used to define the
segregation between tumor and control samples accordingly to their
methylation (beta) values.
}
\examples{
auc_data <- compute_AUC(tumor_toy_table, control_toy_table)
}
