% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_z_score.R
\name{compute_z_score}
\alias{compute_z_score}
\title{Compute a Z-like score to estimate quality of DMRs}
\usage{
compute_z_score(tumor_table, control_table, dmr_table, reference_table,
  out_folder = NULL)
}
\arguments{
\item{tumor_table}{A matrix of beta-values}

\item{control_table}{A matrix of beta-values}

\item{dmr_table}{A data.frame generated by [whole_genome_segmentator]}

\item{reference_table}{A data.frame reporting the genomic coordinates of
each CpG site}

\item{out_folder}{Directory where to store the data generated. Default to
NULL to skip saving}
}
\description{
Compute a Z-like score to estimate quality of DMRs
}
