% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_example}
\alias{control_example}
\title{AUC toy example}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 114 rows and 50 columns.
}
\usage{
data(auc_toy_vector)
}
\description{
Dataset of beta values to run examples (control)
}
\keyword{datasets}
