% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_toy_table}
\alias{control_toy_table}
\title{Control toy table}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 81068 rows and 21 columns.
}
\usage{
data(control_toy_table)
}
\description{
A toy table of beta values (control)
}
\keyword{datasets}
