% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_state_finder.R
\name{fix_na_segments}
\alias{fix_na_segments}
\title{Fix NAs Methylation States}
\usage{
fix_na_segments(meth_states, cutoff)
}
\arguments{
\item{meth_states}{An integer vector of methylation states}

\item{cutoff}{Length cutoff: longer segments will be ignored}
}
\description{
Given a vector of methylation states, this function sets NA stretches
shorter than ore equal to cutoff to the state of flanking sites if flanking
sites have the same state.
}
\keyword{internal}
