% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_state_finder.R
\name{fix_na_states}
\alias{fix_na_states}
\title{Fix NAs Methylation States}
\usage{
fix_na_states(meth_states, cutoff)
}
\arguments{
\item{meth_states}{An integer vector of methylation states}

\item{cutoff}{An integer}
}
\description{
Change NA Methylation States according to flanking sites.
If a stretch of NAs has flanking sites with the same state,
the function changes the NAs to the state of the flaking sites.
If flanking sites are different, if the stretch of NAs is shorter that cutoff,
the function changes the NAs to state of the previous stretch (if first stretch, to the
state of following strech).
}
\keyword{internal}
