% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_short_segments.R
\name{fix_short_segments}
\alias{fix_short_segments}
\title{Fix Short Methylation States}
\usage{
fix_short_segments(meth_states, cutoff)
}
\arguments{
\item{meth_states}{An integer vector of methylation states}

\item{cutoff}{Length cutoff: longer segments will be ignored}
}
\description{
Given a vector of methylation states, this function sets any hyper-(3) or
hypo-(1) methylated stretches of sites shorter than or equal to cutoff, to
non-differential state (2).
}
\keyword{internal}
