% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illumina450k_hg19}
\alias{illumina450k_hg19}
\title{Illumina 450K hg19}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 482421 rows and 5 columns.
}
\usage{
data(illumina450k_hg19)
}
\description{
Reference table for the illumina 450K platform, hg19 coordinates
}
\keyword{datasets}
