% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illumina450k_hg38}
\alias{illumina450k_hg38}
\title{Illumina 450K hg38}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 482421 rows and 10 columns.
}
\usage{
data(illumina450k_hg38)
}
\description{
Reference table for the illumina 450K platform, hg38 coordinates
}
\keyword{datasets}
