% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illuminaEPIC_hg19}
\alias{illuminaEPIC_hg19}
\title{Illumina EPIC hg19}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 865859 rows and 5 columns.
}
\usage{
data(illuminaEPIC_hg19)
}
\description{
Reference table for the illumina EPIC platform, hg19 coordinates
}
\keyword{datasets}
