% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DMR.R
\name{plot_DMR}
\alias{plot_DMR}
\title{Plot information about DMRs}
\usage{
plot_DMR(tumor_table, control_table, auc_vector, genome_version = "hg19")
}
\description{
A way to visualize Differential Methylated Regions showing beta values, 
AUC, and Methylation States
}
