% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRAD_auc}
\alias{PRAD_auc}
\title{PRAD AUC}
\format{
An object of class \code{numeric} of length 482421.
}
\usage{
data(PRAD_auc)
}
\description{
AUC vector obtained with TCGA-PRAD dataset included in the package
}
\keyword{datasets}
