% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRAD_molecular_subtypes}
\alias{PRAD_molecular_subtypes}
\title{PRAD molecular subtypes}
\format{
An object of class \code{data.frame} with 333 rows and 2 columns.
}
\usage{
data(PRAD_molecular_subtypes)
}
\description{
The molecular subtypes annotation of TCGA-PRAD data
}
\keyword{datasets}
