% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRAD_NT_subset}
\alias{PRAD_NT_subset}
\title{PRAD NT subset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 482421 rows and 10 columns.
}
\usage{
data(PRAD_NT_subset)
}
\description{
A subset of normal adjacent samples from the TCGA-PRAD dataset (porstate adenocarcinoma)
}
\keyword{datasets}
