% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRAD_sample_score}
\alias{PRAD_sample_score}
\title{PRAD sample score}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(PRAD_sample_score)
}
\description{
Single sample scores computed on TCGA-PRAD subset
}
\keyword{datasets}
