% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRAD_TP_subset}
\alias{PRAD_TP_subset}
\title{PRAD TP subset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 482421 rows and 20 columns.
}
\usage{
data(PRAD_TP_subset)
}
\description{
A subset of primary tumor samples from the TCGA-PRAD dataset (porstate adenocarcinoma)
}
\keyword{datasets}
