% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_example}
\alias{reference_example}
\title{Reference table example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 114 rows and 3 columns.
}
\usage{
data(reference_example)
}
\description{
Reference table for examples
}
\keyword{datasets}
