% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_toy_table}
\alias{reference_toy_table}
\title{Reference toy table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81068 rows and 5 columns.
}
\usage{
data(reference_toy_table)
}
\description{
Reference table for toy dataset
}
\keyword{datasets}
