% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{Run all ROCker-meth analysis}
\usage{
run_all(beta_table, state, reference_table, nclust = 1, NA_thr = 0.5)
}
\arguments{
\item{nclust}{Number of clusters to use in parallel}

\item{NA_thr}{Fraction of NAs (considered independently in tumor and
control samples) above which a site will not be selected (default=0)}
}
\description{
A simple wrapper to perform a complete analysis with rocker meth. It doesn't
return anything but saves two files for further analysis and evaluation.
}
