% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumor_example}
\alias{tumor_example}
\title{Tumor example}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 114 rows and 50 columns.
}
\usage{
data(tumor_example)
}
\description{
Dataset of beta values to run examples (tumor)
}
\keyword{datasets}
