% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumor_toy_table}
\alias{tumor_toy_table}
\title{Tumor toy table}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 81068 rows and 79 columns.
}
\usage{
data(auc_toy_vector)
}
\description{
A toy table of beta values (tumor)
}
\keyword{datasets}
