% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Generate files from a set of DMR}
\usage{
write_output(dmr_table, sample_score, path, qvalue_thr = 0.05)
}
\arguments{
\item{dmr_table}{A data.frame generated by [whole_genome_segmentator].}

\item{sample_score}{A data.frame generated by [calc_z_score].}

\item{path}{A string including absolute or relative path: "_hyper.bed" and
"_hypo.bed" are automatically added.}

\item{qvalue_thr}{A numerical threshold. Report only DMRs with a q-value
lower than threshold.}
}
\description{
For a quick visualization using a genome browser, this function generates
one BED file (https://bit.ly/1hdVNjA) and two SEG files
(https://bit.ly/2J2WZdZ) reporting statistically significant
hyper-methylated and hypo-methylated DMRs.
}
\details{
BED files coordinates are 0-based while SEG files 1-based.  BED file reports
genomic location and q-values as -log10(q) associated to DMRs.  One SEG file
reports Z-scores for single samples, the other reports average beta
differences (tumor-control) for the entire set of samples (e.g. tumor type).
}
