% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_z_scores.R
\name{write_z_scores}
\alias{write_z_scores}
\title{Generate SEG file from a table of Z-scores}
\usage{
write_z_scores(z_scores_data, path)
}
\arguments{
\item{z_scores_data}{A list generated by [compute_z_scores].}

\item{path}{A string including absolute or relative path which a prefix is added to
("_z_scores.set")}
}
\value{
Nothing. Only saves output tables.
}
\description{
For a quick visualization using a genome browser like IGV, generates
a [SEG](https://bit.ly/2J2WZdZ) file scores reporting Z-scores for tumor and
normal samples generated by [compute_z_scores].
}
\details{
NOTE: SEG files have 1-based coordinates.
}
