% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AUC.R
\name{compute_AUC}
\alias{compute_AUC}
\title{Compute Area Under Curve for a matrix of samples}
\usage{
compute_AUC(tumor_table, control_table, nclust = 1, NA_thr = 0)
}
\arguments{
\item{tumor_table}{A matrix of methylation (beta) values produced by a 450k Illumina
BeadChip from a \bold{tumor} sample}

\item{control_table}{A matrix of methylation (beta) values produced by a 450k Illumina
BeadChip from a \bold{control} sample}

\item{nclust}{Number of clusters to use in parallel}

\item{NA_thr}{Fraction of NAs (considered independently in tumor and
control samples) above which a site will not be selected (default=0)}

\item{sample_state}{A logical vector matching columns in beta_table stating
wheather a sample is a tumor (TRUE) or normal (FALSE)}
}
\value{
A vector of AUC scores
}
\description{
This function compute the Area Under Curve that is then used to define the segregation
between tumor and normal samples accordingly to their methylation (beta) values.
}
