% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_z_score.R
\name{compute_z_score}
\alias{compute_z_score}
\title{Compute a Z-like score to estimate quality of DMRs}
\usage{
compute_z_score(tumor_table, control_table, dmr_table, reference_table)
}
\arguments{
\item{tumor_table}{A matrix of beta-values from tumor samples.}

\item{control_table}{A matrix of beta-values from normal/control samples.}

\item{dmr_table}{A data.frame generated by [whole_genome_segmentator].}

\item{reference_table}{A data.frame reporting the genomic coordinates of
each CpG site.}
}
\description{
Compute a Z-like score to estimate quality of DMRs
}
