% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_z_scores.R
\name{compute_z_scores}
\alias{compute_z_scores}
\title{Compute a Z-like score to estimate quality of DMRs}
\usage{
compute_z_scores(tumor_table, control_table, dmr_table, reference_table,
  min_size)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (fraction) from tumor samples.}

\item{control_table}{A matrix of beta-values (fraction) from normal/control samples.}

\item{dmr_table}{A data.frame generated by [whole_genome_segmentator].}

\item{reference_table}{A data.frame reporting the genomic coordinates of each CpG site in tumor and control matrices.}

\item{min_size}{Minimum number of CpG sites inside DMR. If not reached NAs will be returned.}
}
\description{
Compute a Z-like score to estimate quality of DMRs
}
