% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_z_scores.R
\name{compute_z_scores}
\alias{compute_z_scores}
\title{Compute a Z-like score to estimate quality of DMRs}
\usage{
compute_z_scores(tumor_table, control_table, dmr_table, reference_table,
  min_size = 3)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (fraction) from tumor samples.}

\item{control_table}{A matrix of beta-values (fraction) from normal/control samples.}

\item{dmr_table}{A data.frame reporting the genomic location of DMRs
(chromosome, start, end) (likely produced by [whole_genome_segmentator]).}

\item{reference_table}{A data.frame reporting the genomic coordinates of
each CpG site in tumor and control matrices.}

\item{min_size}{Minimum number of CpG sites inside DMR to compute Z-score
(default = 3, return NA if lower).}
}
\value{
A list of 4 tables: z-scores of DMRs, median beta of DMRs in
tumor samples, median beta of DMRs in normal/control samples and fraction of
NA CpG sites within DMRs.
}
\description{
Assayed CpG may differ from the sites used in auc computation and
segmentation. However, using the same assay for 'Normal' and 'Tumor' is
strongly recommended.
}
