% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_state_finder.R
\name{meth_state_finder}
\alias{meth_state_finder}
\title{Find Methylation Status}
\usage{
meth_state_finder(
  input_signal,
  input_pos,
  auc_sd,
  pt_start,
  normdist,
  ratiosd,
  mu,
  use_trunc
)
}
\arguments{
\item{input_signal}{A numeric vector of AUC scores.}

\item{input_pos}{An integer vector of chromosomal locations}

\item{auc_sd}{Standard deviation of AUC signal (genome wide).}

\item{pt_start}{Transition probability of the HSLM.}

\item{normdist}{Distance normalization parameter of the HSLM.}

\item{ratiosd}{Fraction between the standard deviation of AUC values of
differentially methylated sites and the total standard deviation of AUC
values.}

\item{mu}{Expected mean (AUC) for hypo-methylated state (1-mu is the
expected mean for hyper-methylated state).}

\item{use_trunc}{Use truncated normal distribution (DEBUGGING
ONLY).}
}
\value{
An integer vector of methylation states (1,2,3).
}
\description{
This function classifies CpG sites as hypo-methylated (1), non-differentially
methylated (2) or hyper-methylated (3) using a Heterogeneous Hidden Markov Model (HMM).
}
