% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentator.R
\name{segmentator}
\alias{segmentator}
\title{Define Differentially Methylated Regions}
\usage{
segmentator(meth_states, tumor_beta_mean, control_beta_mean)
}
\arguments{
\item{meth_states}{A vector of methylation states (1, 2, 3, NA).}

\item{tumor_beta_mean}{A vector of average methylation (beta) values.}

\item{control_beta_mean}{A vector of average methylation (beta) values.}
}
\value{
A data.frame with information about segments.
}
\description{
Given a set of methylation states produced by [meth_states_finder] and a set
of methylation (beta) values differences between tumor and control samples,
this function defines stretches of equal states (segments).
}
\keyword{internal}
