% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_genome_segmentator.R
\name{whole_genome_segmentator}
\alias{whole_genome_segmentator}
\title{Find Differentially Methylated Regions across whole genome}
\usage{
whole_genome_segmentator(tumor_table, control_table, auc_vector,
  reference_table, na_cutoff = 5, length_cutoff = 3)
}
\arguments{
\item{tumor_table}{A matrix of beta values}

\item{control_table}{A matrix of beta values}

\item{auc_vector}{An numeric vector generated by [compute_AUC]}

\item{reference_table}{A data.frame reporting location of CpG probes (column and genomic_position)}

\item{na_cutoff}{}

\item{length_cutoff}{}
}
\description{
A function to define DMR for an entire chromosome
}
