% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_genome_segmentator.R
\name{whole_genome_segmentator}
\alias{whole_genome_segmentator}
\title{Find Differentially Methylated Regions across whole genome}
\usage{
whole_genome_segmentator(tumor_table, control_table, auc_vector,
  reference_table, length_cutoff = 5, na_cutoff = 0)
}
\arguments{
\item{tumor_table}{A matrix of beta values}

\item{control_table}{A matrix of beta values}

\item{auc_vector}{An numeric vector generated by [compute_AUC]}

\item{reference_table}{A data.frame reporting location of CpG probes (chromosome
and genomic_position)}

\item{length_cutoff}{An integer to remove streches of differential
methylation shorter than cutoff}

\item{na_cutoff}{An integer to set stretches of NAs to the state of flanking sites
depending if the stretches are shorter than cutoff}
}
\description{
A function to define DMR for an entire genome
}
