% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_genome_segmentator.R
\name{whole_genome_segmentator}
\alias{whole_genome_segmentator}
\title{Find Differentially Methylated Regions across whole genome}
\usage{
whole_genome_segmentator(tumor_table, control_table, auc_vector,
  reference_table, length_cutoff = 5, pt_start = 0.05,
  normdist = 1e+05, ratiosd = 0.4)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (percentage) from tumor samples.}

\item{control_table}{A matrix of beta-values (percentage) from
normal/control samples.}

\item{auc_vector}{A numeric vector generated by [compute_AUC].}

\item{reference_table}{A data.frame reporting the location of CpG sites
(chromosome and genomic_position).}

\item{length_cutoff}{An integer to remove streches of differential
methylation shorter than cutoff.}

\item{pt_start}{Transition probability of the HSLM. Default is 0.05.}

\item{normdist}{Distance normalization parameter of the HSLM. Default is 1e5.}

\item{ratiosd}{Fraction between the standard deviation of AUC values of
differentially methylated sites and the total standard deviation of AUC
values. Default is 0.4.}
}
\description{
A function to define DMR for an entire genome
}
