% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_genome_segmentator.R
\name{whole_genome_segmentator}
\alias{whole_genome_segmentator}
\title{Find Differentially Methylated Regions across whole genome}
\usage{
whole_genome_segmentator(tumor_table, control_table, auc_vector,
  reference_table, length_cutoff = 5, pt_start = 0.05,
  normdist = 100000, ratiosd = 0.4, mu = 0.25, use_trunc = TRUE)
}
\arguments{
\item{tumor_table}{A matrix of beta-values (percentage) from tumor samples.}

\item{control_table}{A matrix of beta-values (percentage) from
normal/control samples.}

\item{auc_vector}{A numeric vector generated by [compute_AUC].}

\item{reference_table}{A data.frame reporting the location of CpG sites
(chromosome and genomic_position).}

\item{length_cutoff}{An integer to remove streches of differential
methylation shorter than cutoff.}

\item{pt_start}{Transition probability of the HSLM. Default is 0.05.}

\item{normdist}{Distance normalization parameter of the HSLM. Default is 1e5.}

\item{ratiosd}{Fraction between the standard deviation of AUC values of
differentially methylated sites and the total standard deviation of AUC
values. Default is 0.4.}

\item{mu}{Expected mean (AUC) for hypo-methylated state (1-mu is the
expected mean for hyper-methylated state). Default is 0.25.}

\item{use_trunc}{Use truncated normal distribution (DEBUGGING ONLY).
Default is TRUE.}
}
\value{
A data.frame reporting genomic location, number of CpG sites,
methylation state, average beta difference (tumor vs. control), p-value and
adjusted (Benjamini-Hochberg) p-value (q-value) of discovered DMRs.
}
\description{
A function to define DMR for an entire genome.
}
