% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bed.R
\name{write_bed}
\alias{write_bed}
\title{Generate a bed file from a set of DMR}
\usage{
write_bed(dmr_table, path, qvalue_thr = 0.05, one_based = TRUE)
}
\arguments{
\item{dmr_table}{A data.frame generated by [whole_genome_segmentator].}

\item{path}{A character vector containing a single path name.}

\item{qvalue_thr}{A numerical threshold. Report only DMRs with a q-value
lower than threshold.}

\item{one_based}{A logical. Wether the output coordinates are 0-bases or
1-based (default).}
}
\description{
This function generates a [bed format file](https://bit.ly/1hdVNjA)
for a quick visualization using a genome browser.
}
