% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Generate files from a set of DMR}
\usage{
write_output(path, dmr_table, sample_score = NULL, qvalue_thr = 0.05)
}
\arguments{
\item{path}{A string including absolute or relative path: "_hyper.bed" and
"_hypo.bed" are automatically added.}

\item{dmr_table}{A data.frame generated by [whole_genome_segmentator].}

\item{sample_score}{A list generated by [compute_z_score].}

\item{qvalue_thr}{A numerical threshold. Report only DMRs with a q-value
lower than threshold.}
}
\description{
For a quick visualization using a genome browser like IGV, this function
generates [BED](https://bit.ly/1hdVNjA) and [SEG](https://bit.ly/2J2WZdZ)
files of statistically significant hyper-methylated and hypo-methylated
DMRs.
}
\details{
BED files (hyper- and hypo-methylated DMRs separately) report genomic
location and q-value (in form of -log10(q)).
SEG files report average beta differences between tumor and normal/control
samples at each DMR.
Additionally, a SEG file reporting Z-scores for each tumor sample is
generated providing a list produced by [compute_z_score].

NOTE: BED files have 0-based coordinates while SEG files have 1-based
coordinates.
}
